Added a GameScripting system where you can assign GameScript objects to Replica Objects in the world to early manipulate them.
Added almost full functionality to Quickbuilds.
Known flaws: Imagination is not drawn if the quickbuild is started in a incomplete state; numbers may not be correct.

Made the stats on the player more useable. You can now pick up coins, hearts, and armor pickups and have them add to your stats.

Added a /give chat command.

Added a handler for moving stuff around in the inventory.

Plus more stuff I can't remember of the top of my head.

Signed-off-by: wincent01 <wincent.holm@gmail.com>